requirements.txt
git+https://github.com/openai/whisper.git
git+https://github.com/huggingface/datasets.git
git+https://github.com/huggingface/transformers.git
setuptools-rust
librosa
numpy==1.26.4
soundfile
resampy==0.4.3
pydub
SpeechRecognition==3.10.4
pyAudioAnalysis
tensorflow
scikit-learn
pandas
hmmlearn
imblearn
streamlit
plotly
eyed3
feat
pyyaml
tkinter
faiss-cpu
torch